added not found checks in admin controller and changed the message on errors, checking for problems encouraged.

+++ controllers/admin/achievements/achievementController.js
@@ -14,20 +14,22 @@ const createAchievement = async (req, res) => {
   const data = await newAchievement.createAchievement();
 
   if (!data) {
-    throw new BadRequestError(`Error in creating achievement. Try again later.`);
+    throw new BadRequestError(`There was a problem in creating the achievement ${name}.`);
   }
 
   const users = await User.getRawUsers();
 
   if (!users) {
-    throw new BadRequestError(`Error in getting users.`);
+    throw new BadRequestError(
+      `There was a problem in getting the users for assigning the achievement.`
+    );
   }
 
   users?.map(async (user) => {
     const assignAchievement = await UserAchievement.assignAchievement(data.insertId, user.user_id);
 
     if (!assignAchievement) {
-      throw new BadRequestError(`Error in assigning achievement to users.`);
+      throw new BadRequestError(`There was a problem in assigning the achievement to the users.`);
     }
   });
 
@@ -39,6 +41,12 @@ const updateAchievement = async (req, res) => {
   const { achievement_id } = req.params;
   const { achievement } = req.body;
   const { achievement_name, achievement_type, task, specifics, goal, reward_id } = achievement;
+
+  const ifExist = await Achievement.getAchievement(achievement_id);
+
+  if (!ifExist) {
+    throw new BadRequestError(`The achievement you are trying to update does not exist.`);
+  }
 
   const data = await Achievement.updateAchievement(
     achievement_id,
@@ -52,7 +60,9 @@ const updateAchievement = async (req, res) => {
   );
 
   if (!data) {
-    throw new BadRequestError(`Error in updating achievement. Try again later.`);
+    throw new BadRequestError(
+      `There was a problem in updating the achievement ${achievement_name}.`
+    );
   }
 
   res.status(StatusCodes.OK).json(data);
@@ -61,10 +71,16 @@ const updateAchievement = async (req, res) => {
 const deleteAchievement = async (req, res) => {
   const { achievement_id } = req.params;
 
+  const ifExist = await Achievement.getAchievement(achievement_id);
+
+  if (!ifExist) {
+    throw new BadRequestError(`The achievement you are trying to delete does not exist.`);
+  }
+
   const achievement = await Achievement.deleteAchievement(achievement_id);
 
   if (!achievement) {
-    throw new BadRequestError(`Error in deleting achievement. Try again later.`);
+    throw new BadRequestError(`There was a problem in deleting the achievement.`);
   }
 
   res.status(StatusCodes.OK).json(achievement);
@@ -72,6 +88,7 @@ const deleteAchievement = async (req, res) => {
 
 const getAllAchievements = async (req, res) => {
   const { searchFilter, goalRangeFilter, sortFilter, dateRangeFilter } = req.query;
+
   const achievement = await Achievement.getAllAchievements(
     searchFilter,
     goalRangeFilter,
@@ -80,7 +97,7 @@ const getAllAchievements = async (req, res) => {
   );
 
   if (!achievement) {
-    throw new BadRequestError(`Error in getting all achievements. Try again later.`);
+    throw new BadRequestError(`There was a problem in getting the achievements.`);
   }
 
   res.status(StatusCodes.OK).json(achievement);
@@ -92,7 +109,7 @@ const getAchievement = async (req, res) => {
   const achievement = await Achievement.getAchievement(achievement_id);
 
   if (!achievement) {
-    throw new BadRequestError(`Error in getting achievement. Try again later.`);
+    throw new BadRequestError(`The achievement you want to view does not exist.`);
   }
 
   res.status(StatusCodes.OK).json(achievement);
