+++ controllers/admin/achievements/rewardController.js
@@ -11,7 +11,7 @@ const createReward = async (req, res) => {
   const data = await newReward.createReward();
 
   if (!data) {
-    throw new BadRequestError(`Error in creating reward. Try again later.`);
+    throw new BadRequestError(`There was a problem in creating the reward ${name}.`);
   }
 
   res.status(StatusCodes.OK).json(data);
@@ -21,6 +21,12 @@ const updateReward = async (req, res) => {
   const { id } = req.user;
   const { reward_id } = req.params;
   const { reward_name, reward_type, description, reward } = req.body;
+
+  const ifExist = await Reward.getReward(reward_id);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The reward you are trying to update does not exist.`);
+  }
 
   const data = await Reward.updateReward(
     reward_id,
@@ -32,7 +38,7 @@ const updateReward = async (req, res) => {
   );
 
   if (!data) {
-    throw new BadRequestError(`Error in updating reward. Try again later.`);
+    throw new BadRequestError(`There was a problem in updating the reward ${reward_name}.`);
   }
 
   res.status(StatusCodes.OK).json(data);
@@ -41,10 +47,16 @@ const updateReward = async (req, res) => {
 const deleteReward = async (req, res) => {
   const { reward_id } = req.params;
 
+  const ifExist = await Reward.getReward(reward_id);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The reward you are trying to delete does not exist.`);
+  }
+
   const data = await Reward.deleteReward(reward_id);
 
   if (!data) {
-    throw new BadRequestError(`Error in deleting reward. Try again later.`);
+    throw new BadRequestError(`There was a problem in deleting the reward.`);
   }
 
   res.status(StatusCodes.OK).json(data);
@@ -55,7 +67,7 @@ const getAllRewards = async (req, res) => {
   const data = await Reward.getAllRewards(searchFilter, sortFilter, dateRangeFilter, typeFilter);
 
   if (!data) {
-    throw new BadRequestError(`Error in getting all rewards. Try again later.`);
+    throw new BadRequestError(`There was a problem in getting all the rewards.`);
   }
 
   res.status(StatusCodes.OK).json(data);
@@ -64,10 +76,16 @@ const getAllRewards = async (req, res) => {
 const getReward = async (req, res) => {
   const { reward_id } = req.params;
 
+  const ifExist = await Reward.getReward(reward_id);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The reward you are trying to view does not exist.`);
+  }
+
   const data = await Reward.getReward(reward_id);
 
   if (!data) {
-    throw new BadRequestError(`Error in getting reward. Try again later.`);
+    throw new BadRequestError(`There was a problem in getting the reward.`);
   }
 
   res.status(StatusCodes.OK).json(data);
