+++ controllers/admin/minigames/riddlesController.js
@@ -11,7 +11,7 @@ const createRiddle = async (req, res) => {
   const data = await riddles.createRiddle();
 
   if (!data) {
-    throw new BadRequestError(`Error in creating riddle. Try again later.`);
+    throw new BadRequestError(`There was a problem in creating the riddle ${answer}.`);
   }
 
   res.status(StatusCodes.OK).json(data);
@@ -22,10 +22,16 @@ const updateRiddle = async (req, res) => {
   const { riddle_id } = req.params;
   const { riddle, answer } = req.body;
 
+  const ifExist = await Riddles.getRiddle(riddle_id);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The riddle you are trying to update does not exist.`);
+  }
+
   const riddles = await Riddles.updateRiddle(riddle_id, riddle, answer, id);
 
   if (!riddles) {
-    throw new BadRequestError(`Error in updating riddle. Try again later.`);
+    throw new BadRequestError(`There was a problem in updating the riddle ${answer}.`);
   }
 
   res.status(StatusCodes.OK).json(riddles);
@@ -34,10 +40,16 @@ const updateRiddle = async (req, res) => {
 const deleteRiddle = async (req, res) => {
   const { riddle_id } = req.params;
 
+  const ifExist = await Riddles.getRiddle(riddle_id);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The riddle you are trying to update does not exist.`);
+  }
+
   const riddles = await Riddles.deleteRiddle(riddle_id);
 
   if (!riddles) {
-    throw new BadRequestError(`Error in deleting riddle. Try again later.`);
+    throw new BadRequestError(`There was a problem in deleting the riddle.`);
   }
 
   res.status(StatusCodes.OK).json(riddles);
@@ -48,7 +60,7 @@ const getAllRiddles = async (req, res) => {
   const riddles = await Riddles.getAllRiddles(searchFilter, sortFilter, dateRangeFilter);
 
   if (!riddles) {
-    throw new BadRequestError(`Error in deleting riddle. Try again later.`);
+    throw new BadRequestError(`There was a problem in getting all riddles.`);
   }
 
   res.status(StatusCodes.OK).json(riddles);
@@ -60,7 +72,7 @@ const getRiddle = async (req, res) => {
   const riddles = await Riddles.getRiddle(riddle_id);
 
   if (!riddles) {
-    throw new BadRequestError(`Error in deleting riddle. Try again later.`);
+    throw new NotFoundError(`The riddle you are trying to view does not exist.`);
   }
 
   res.status(StatusCodes.OK).json(riddles);
@@ -70,7 +82,7 @@ const getRandomRiddle = async (req, res) => {
   const riddle = await Riddles.getRandomRiddle();
 
   if (!riddle) {
-    throw new BadRequestError(`Error in getting random riddle. Try again later.`);
+    throw new BadRequestError(`There was a problem in getting a random riddle.`);
   }
 
   res.status(StatusCodes.OK).json(riddle);
