+++ controllers/admin/story/storyContentController.js
@@ -1,6 +1,7 @@
 const { StatusCodes } = require("http-status-codes");
 const { BadRequestError, NotFoundError } = require("../../../errors");
 const StoryContent = require("../../../models/story/StoryContent");
+const Story = require("../../../models/story/Story");
 
 const createContent = async (req, res) => {
   const { id } = req.user;
@@ -11,7 +12,7 @@ const createContent = async (req, res) => {
   const data = await storyContent.createContent();
 
   if (!data) {
-    throw new BadRequestError(`Error in creating content. Try again later.`);
+    throw new BadRequestError(`There was a problem in creating the content.`);
   }
 
   res.status(StatusCodes.OK).json(data);
@@ -22,20 +23,32 @@ const updateContent = async (req, res) => {
   const { id } = req.user;
   const { page, content, image } = req.body;
 
+  const ifExist = await StoryContent.getContent(content_id);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The story content you are trying to update does not exist.`);
+  }
+
   const storyContent = await StoryContent.updateContent(content_id, page, content, image, id);
 
   if (!storyContent) {
-    throw new BadRequestError(`Error in updating content. Try again later.`);
+    throw new BadRequestError(`There was a problem in updating the story content.`);
   }
 };
 
 const deleteContent = async (req, res) => {
   const { content_id } = req.params;
 
+  const ifExist = await StoryContent.getContent(content_id);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The story content you are trying to delete does not exist.`);
+  }
+
   const storyContent = await StoryContent.deleteContent(content_id);
 
   if (!storyContent) {
-    throw new BadRequestError(`Error in deleting content. Try again later.`);
+    throw new BadRequestError(`There was a problem in deleting the story content.`);
   }
 
   res.status(StatusCodes.OK).json(storyContent);
@@ -44,10 +57,16 @@ const deleteContent = async (req, res) => {
 const getAllContent = async (req, res) => {
   const { storyId } = req.query;
 
+  const ifExist = await Story.getStory(storyId);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The story you are trying to view does not exist.`);
+  }
+
   const storyContent = await StoryContent.getAllContent(storyId);
 
   if (!storyContent) {
-    throw new BadRequestError(`Error in getting all content. Try again later.`);
+    throw new BadRequestError(`There was a problem in getting all content.`);
   }
 
   res.status(StatusCodes.OK).json(storyContent);
@@ -59,7 +78,7 @@ const getContent = async (req, res) => {
   const storyContent = await StoryContent.getContent(content_id);
 
   if (!storyContent) {
-    throw new BadRequestError(`Error in getting all content. Try again later.`);
+    throw new NotFoundError(`The story content you are trying to view does not exist.`);
   }
 
   res.status(StatusCodes.OK).json(storyContent);
