+++ controllers/admin/story/storyController.js
@@ -13,7 +13,7 @@ const createStory = async (req, res) => {
   const data = await story.createStory();
 
   if (!data) {
-    throw new BadRequestError(`Error in creating story. Try again later.`);
+    throw new BadRequestError(`There was a problem in creating the story ${title}.`);
   }
 
   pages.map(async (page) => {
@@ -29,7 +29,9 @@ const createStory = async (req, res) => {
     const newPage = await pageContent.createContent();
 
     if (!newPage) {
-      throw new BadRequestError(`Error in adding page content. Try again later.`);
+      throw new BadRequestError(
+        `There was a problem in adding the content in page ${page.pageNumber}.`
+      );
     }
   });
 
@@ -42,19 +44,25 @@ const updateStory = async (req, res) => {
   const { story_id } = req.params;
   const { id } = req.user;
 
+  const ifExist = await Story.getStory(story_id);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The story you are trying to update does not exist.`);
+  }
+
   const bookCover = file?.src ? file?.src : book_cover ? book_cover : null;
 
   const data = await Story.updateStory(story_id, title, author, bookCover, lexile, genre, id);
 
   if (!data) {
-    throw new BadRequestError(`Error in creating story. Try again later.`);
+    throw new BadRequestError(`There was a problem in updating the story ${title}.`);
   }
 
   toDelete.map(async (contentId) => {
     const deleteContent = await StoryContent.deleteContent(contentId);
 
     if (!deleteContent) {
-      throw new BadRequestError(`Error in deleting story content. Try again later.`);
+      throw new BadRequestError(`There was a problem in deleting the story content.`);
     }
   });
 
@@ -68,9 +76,17 @@ const updateStory = async (req, res) => {
       const addPage = await newPage.createContent();
 
       if (!addPage) {
-        throw new BadRequestError(`Error in adding page content. Try again later.`);
+        throw new BadRequestError(
+          `There was a problem in adding the new page ${page.page} content.`
+        );
       }
     } else {
+      const ifExist = await StoryContent.getContent(content_id);
+
+      if (!ifExist) {
+        throw new NotFoundError(`The story content you are trying to update does not exist.`);
+      }
+
       const editPage = await StoryContent.updateContent(
         content_id,
         page.page,
@@ -92,10 +108,16 @@ const updateStory = async (req, res) => {
 const deleteStory = async (req, res) => {
   const { story_id } = req.params;
 
+  const ifExist = await Story.getStory(story_id);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The story you are trying to delete does not exist.`);
+  }
+
   const story = await Story.deleteStory(story_id);
 
   if (!story) {
-    throw new BadRequestError(`Error in deleting story. Try again later.`);
+    throw new BadRequestError(`There was a problem in deleting the story.`);
   }
 
   res.status(StatusCodes.OK).json(story);
@@ -111,7 +133,7 @@ const getAllStories = async (req, res) => {
   );
 
   if (!story) {
-    throw new BadRequestError(`Error in getting all stories. Try again later.`);
+    throw new BadRequestError(`There was a problem in getting all the stories.`);
   }
 
   res.status(StatusCodes.OK).json(story);
@@ -123,7 +145,7 @@ const getStory = async (req, res) => {
   const story = await Story.getStory(story_id);
 
   if (!story) {
-    throw new BadRequestError(`Error in getting story. Try again later.`);
+    throw new NotFoundError(`The story you are trying to view does not exist.`);
   }
 
   res.status(StatusCodes.OK).json(story);
