+++ controllers/admin/test/testQuestionController.js
@@ -1,17 +1,24 @@
 const { StatusCodes } = require("http-status-codes");
 const { BadRequestError, NotFoundError } = require("../../../errors");
 const TestQuestion = require("../../../models/test/TestQuestion");
+const Test = require("../../../models/test/Test");
 
 const createQuestion = async (req, res) => {
   const { id } = req.user;
   const { test_id, question } = req.body;
+
+  const ifExist = await Test.getTest(test_id);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The test you are trying to add a question to does not exist.`);
+  }
 
   const testQuestion = new TestQuestion(test_id, question, id);
 
   const data = await testQuestion.createQuestion();
 
   if (!data) {
-    throw new BadRequestError(`Error in creating question. Try again later.`);
+    throw new BadRequestError(`There was a problem in creating the test question.`);
   }
 
   res.status(StatusCodes.OK).json(data);
@@ -22,10 +29,16 @@ const updateQuestion = async (req, res) => {
   const { id } = req.user;
   const { question } = req.body;
 
+  const ifExist = await TestQuestion.getQuestion(question_id);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The test question you are updating does not exist.`);
+  }
+
   const testQuestion = await TestQuestion.updateQuestion(question_id, question, id);
 
   if (!testQuestion) {
-    throw new BadRequestError(`Error in updating question. Try again later.`);
+    throw new BadRequestError(`There was a problem in updating the test question.`);
   }
 
   res.status(StatusCodes.OK).json(testQuestion);
@@ -34,10 +47,16 @@ const updateQuestion = async (req, res) => {
 const deleteQuestion = async (req, res) => {
   const { question_id } = req.params;
 
+  const ifExist = await TestQuestion.getQuestion(question_id);
+
+  if (!ifExist) {
+    throw new NotFoundError(`The test question you are deleting does not exist.`);
+  }
+
   const testQuestion = await TestQuestion.deleteQuestion(question_id);
 
   if (!testQuestion) {
-    throw new BadRequestError(`Error in deleting test question. Try again later.`);
+    throw new BadRequestError(`There was a problem in deleting the test question.`);
   }
 
   res.status(StatusCodes.OK).json(testQuestion);
@@ -49,7 +68,7 @@ const getAllQuestions = async (req, res) => {
   const testQuestion = await TestQuestion.getAllQuestions(testId);
 
   if (!testQuestion) {
-    throw new BadRequestError(`Error in getting all test question. Try again later.`);
+    throw new BadRequestError(`There was a problem in getting all the test questions.`);
   }
 
   res.status(StatusCodes.OK).json(testQuestion);
@@ -61,7 +80,7 @@ const getQuestion = async (req, res) => {
   const testQuestion = await TestQuestion.getQuestion(question_id);
 
   if (!testQuestion) {
-    throw new BadRequestError(`Error in getting test question. Try again later.`);
+    throw new NotFoundError(`The test question you are trying to get does not exist.`);
   }
 
   res.status(StatusCodes.OK).json(testQuestion);
